// Java/Tamin-Projects/eservice-lite/ui/src/app/models/pensioner/pensionRequest.model.ts
export class PensionRequestModel {
  id: string;
  nationalCode: string;
  mobileNumber: string;
  phoneNumber: string;
  issuePlace: string;
  insuranceNumber: string;
  firstName: string;
  lastName: string;
  birthDate: Date;
  fatherName: string;
  address: string;
  idNumber: string;
  workshopName: string;
  workshopNumber: string;
  workshopManager: string;
  activityName: string;
  workshopAddress: string;
  province: string;
  city: string;
  branch: string;
  gender: string;
  age: number;
  operation: string;
  comment: string;
}